refactor: consolidate shared gradient logic
